+++ vite.config.js
@@ -1,7 +1,7 @@
 import path from "path";
 import { defineConfig } from "vite";
 import { svelte } from "@sveltejs/vite-plugin-svelte";
-import { dependencies ,peerDependencies, version } from "./package.json"
+import { dependencies, peerDependencies, version } from "./package.json"
 
 
 export default defineConfig(({ mode }) => {
@@ -10,6 +10,7 @@ export default defineConfig(({ mode }) => {
     return {
         define: {
             __APP_VERSION__: JSON.stringify(version),
+            __APP_DEPENDENCIES__: JSON.stringify(dependencies),
         },
         plugins: [
             svelte({
@@ -23,7 +24,7 @@ export default defineConfig(({ mode }) => {
             lib: {
                 // eslint-disable-next-line no-undef
                 entry: path.resolve(__dirname, "src/index.ts"),
-                formats: ["es","umd"], // TODO cjs 版本待修复
+                formats: ["es", "umd"], // TODO cjs 版本待修复
                 name: "WindowManager",
                 fileName: "index"
             },